% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomap.R
\name{bubblemap}
\alias{bubblemap}
\title{`bubblemap()` prints a world map based on shape files
from the World Health Organization. It requires ggplot2.}
\usage{
bubblemap(
  X = data.frame(iso3 = NA, size = NA),
  line.col = "black",
  map.title = "",
  legend.title = "",
  water.col = "white",
  bubble.col = "orange",
  bubble.alpha = 0.4,
  scale.breaks = waiver(),
  scale.limits = NULL,
  scale.labels = waiver(),
  legend.pos = c(0.14, 0.26),
  recentre = 12
)
}
\arguments{
\item{X}{a dataframe. It must contain a variable "iso" (factor)
with standard WHO ISO3 country codes and a variable "size" (bubble sizes).}

\item{line.col}{Colour of country border lines.}

\item{map.title}{Map title.}

\item{legend.title}{Legend title.}

\item{water.col}{Colour of oceans and lakes.}

\item{bubble.col}{Colour of bubbles.}

\item{bubble.alpha}{Bubble transparency.}

\item{scale.breaks}{Bubble scale breaks.}

\item{scale.limits}{Scale limits.}

\item{scale.labels}{Scale labels.}

\item{legend.pos}{A vector of two numbers, positions the legend.}

\item{recentre}{A longitude value between -180 and 180 set at the middle of the map.}
}
\value{
A ggplot2 plot.
}
\description{
`bubblemap()` prints a world map based on shape files
from the World Health Organization. It requires ggplot2.
}
\examples{
bubblemap(data.frame(iso3=c('BRA','ZAF','RUS','IND','CHN'), size=runif(5)),
         scale.breaks=c(.1,.5),
         legend.title = 'Silly Example')

}
\author{
Philippe Glaziou, adapted from scripts from Tom Hiatt and Hazim Timimi.
}
