% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomap.R
\name{.hidef}
\alias{.hidef}
\title{Choropleth world maps}
\usage{
.hidef(
  X,
  colours,
  low.col,
  high.col,
  line.col,
  line.width,
  map.title,
  legend.title,
  water.col,
  na.label,
  na.col,
  disclaimer,
  legend.pos,
  recentre
)
}
\value{
A ggplot2 plot.
}
\description{
`.hidef()` prints a choropleth world map based on shape files
from the World Health Organization. It requires ggplot2.
}
\author{
Philippe Glaziou
}
