% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomap.R
\name{.hidefbb}
\alias{.hidefbb}
\title{`.hidefbb()` prints a world map based on shape files
from the World Health Organization. It requires ggplot2.}
\usage{
.hidefbb(
  X,
  line.col,
  map.title,
  legend.title,
  water.col,
  bubble.col,
  bubble.alpha,
  scale.breaks,
  scale.limits,
  scale.labels,
  line.width,
  legend.pos,
  recentre
)
}
\value{
A ggplot2 plot.
}
\description{
`.hidefbb()` prints a world map based on shape files
from the World Health Organization. It requires ggplot2.
}
\author{
Philippe Glaziou
}
