% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomap.R
\name{.lodefbb}
\alias{.lodefbb}
\title{`.lodefbb()` prints a world map based on shape files
from the World Health Organization. It requires ggplot2.}
\usage{
.lodefbb(
  X,
  line.col,
  map.title,
  legend.title,
  water.col,
  bubble.col,
  bubble.alpha,
  scale.breaks,
  scale.limits,
  scale.labels,
  legend.pos,
  recentre
)
}
\value{
A ggplot2 plot.
}
\description{
`.lodefbb()` prints a world map based on shape files
from the World Health Organization. It requires ggplot2.
}
\author{
Philippe Glaziou, adapted from scripts from Tom Hiatt and Hazim Timimi.
}
