% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomap.R
\name{whomap}
\alias{whomap}
\title{Choropleth world map}
\usage{
whomap(
  X = data.frame(iso3 = NA, var = NA),
  colours = NULL,
  low.col = "#BDD7E7",
  high.col = "#08519C",
  line.col = "black",
  map.title = "",
  legend.title = "",
  water.col = "white",
  na.label = "No data",
  na.col = "grey95",
  disclaimer = FALSE,
  legend.pos = c(0.09, 0.26),
  recentre = 12
)
}
\arguments{
\item{X}{a dataframe. It must contain a variable "iso" (factor)
with standard WHO ISO3 country codes.The categorical variable to be
mapped should be named "var" (see examples).}

\item{colours}{A vector of colour values for each category in "var", excepting missing values.}

\item{low.col}{First value of a gradient of colours.}

\item{high.col}{Last value of a gradient of colours.}

\item{line.col}{Colour of country border lines.}

\item{map.title}{Map title.}

\item{legend.title}{Legend title.}

\item{water.col}{Colour of oceans and lakes.}

\item{na.label}{Legend lable for missing values.}

\item{na.col}{Colour of countries with missing values.}

\item{disclaimer}{A boolean, inserts a standard WHO disclaimer.}

\item{legend.pos}{A vector of two numbers, positions the legend.}

\item{recentre}{A longitude value between -180 and 180 set at the middle of the map.}
}
\value{
A ggplot2 plot.
}
\description{
`whomap()` prints a choropleth world map based on shape files
from the World Health Organization. It requires ggplot2.
}
\examples{
# A simple world map
whomap()

# Map of the BRICS countries in red, no legend
brics <- data.frame(iso3=c('BRA','CHN','IND','RUS','ZAF'), var=1)
whomap(X=brics, colours=c('red'), legend.pos='none')

# with legend
brics <- data.frame(iso3=c('BRA','CHN','IND','RUS','ZAF'), var=1:5)
whomap(brics, legend.title='BRICS')

}
\author{
Philippe Glaziou, adapted from scripts from Tom Hiatt and Hazim Timimi.
}
